State Dictionary saving/loading (#102)
* adopt code for checkpoint
